chore: generic useEventTracker
