+++ frontend/src/component/providers/LogRocketProvider/LogRocketRunner.tsx
@@ -1,14 +1,20 @@
-import { useEffect } from 'react';
+import { useEffect, useRef } from 'react';
 import LogRocket from 'logrocket';
 import { scrubUrl, scrubBrowserUrl, isStaticAsset } from './scrubUrl';
+import type { LogRocketInstance } from 'contexts/LogRocketContext';
 
 type Props = {
     appId: string;
     clientId: string;
     userId: number;
+    onReady: (instance: LogRocketInstance) => void;
 };
 
-const LogRocketRunner = ({ appId, clientId, userId }: Props) => {
+const LogRocketRunner = ({ appId, clientId, userId, onReady }: Props) => {
+    // make sure we won't re-initialize even if onReady isn't stable
+    const onReadyRef = useRef(onReady);
+    onReadyRef.current = onReady;
+
     useEffect(() => {
         try {
             LogRocket.init(appId, {
@@ -41,6 +47,7 @@ const LogRocketRunner = ({ appId, clientId, userId }: Props) => {
                 clientId,
                 userId: String(userId),
             });
+            onReadyRef.current({ track: LogRocket.track.bind(LogRocket) });
         } catch (error) {
             console.warn(error);
         }
