+++ frontend/src/hooks/usePlausibleTracker.ts
@@ -1,97 +1,8 @@
 import { useCallback, useContext, useRef } from 'react';
 import { PlausibleContext } from 'contexts/PlausibleContext';
-import type { EventOptions, PlausibleOptions } from 'plausible-tracker';
-
-/**
- * Allowed event names. Makes it easy to remove, since TS will complain.
- * Add those to Plausible as Custom event goals.
- * @see https://plausible.io/docs/custom-event-goals#2-create-a-custom-event-goal-in-your-plausible-analytics-account
- * @example `'download | 'invite' | 'signup'`
- **/
-export type CustomEvents =
-    | 'invite'
-    | 'upgrade_plan_clicked'
-    | 'read_about'
-    | 'change_request'
-    | 'favorite'
-    | 'maintenance'
-    | 'banner'
-    | 'hidden_environment'
-    | 'project_overview'
-    | 'suggest_tags'
-    | 'unknown_ui_error'
-    | 'export_import'
-    | 'project_api_tokens'
-    | 'project_stickiness_set'
-    | 'notifications'
-    | 'batch_operations'
-    | 'strategyTitle'
-    | 'default_strategy'
-    | 'demo'
-    | 'demo-start'
-    | 'demo-close'
-    | 'demo-finish'
-    | 'demo-see-plans'
-    | 'demo-see-plan'
-    | 'demo-restart'
-    | 'demo-view-demo-link'
-    | 'demo-start-topic'
-    | 'demo-ask-questions'
-    | 'demo-open-demo-web'
-    | 'demo-open-walkthrough-guide'
-    | 'context-usage'
-    | 'segment-usage'
-    | 'strategy-add'
-    | 'edit-milestone-strategy'
-    | 'suggestion-strategy-add'
-    | 'playground'
-    | 'feature-type-edit'
-    | 'strategy-variants'
-    | 'search-filter-suggestions'
-    | 'project-metrics'
-    | 'open-integration'
-    | 'feature-naming-pattern'
-    | 'project-mode'
-    | 'dependent_features'
-    | 'playground_token_input_used'
-    | 'search-filter'
-    | 'search-feature-buttons'
-    | 'new-strategy-form'
-    | 'feedback'
-    | 'feature-metrics'
-    | 'search-bar'
-    | 'sdk-reporting'
-    | 'insights-share'
-    | 'many-strategies'
-    | 'sdk-banner'
-    | 'feature-lifecycle'
-    | 'command-bar'
-    | 'new-in-unleash-click'
-    | 'new-in-unleash-dismiss'
-    | 'search-opened'
-    | 'events-exported'
-    | 'event-timeline'
-    | 'onboarding'
-    | 'personal-dashboard'
-    | 'order-environments'
-    | 'project-navigation'
-    | 'productivity-report'
-    | 'release-management'
-    | 'feature-links'
-    | 'project-cleanup'
-    | 'project-list-view-toggle'
-    | 'impact-metrics'
-    | 'impact-metrics-safeguards-splash'
-    | 'release-management-splash'
-    | 'upgrade_trial_dialog'
-    | 'upgrade_trial_billing_page'
-    | 'new-template-from-add-strategy'
-    | 'flagpage-impact-metrics'
-    | 'signup-dialog'
-    | 'signup-dialog-error'
-    | 'safeguards'
-    | 'remote-mcp'
-    | 'external-impact-metrics';
+import type { CustomEvents } from 'utils/trackingEvents';
+import type { TrackEventOptions } from 'contexts/EventTrackerContext';
+export type { CustomEvents } from 'utils/trackingEvents';
 
 export const usePlausibleTracker = () => {
     const plausible = useContext(PlausibleContext);
@@ -99,18 +10,8 @@ export const usePlausibleTracker = () => {
     plausibleRef.current = plausible;
 
     const trackEvent = useCallback(
-        (
-            eventName: CustomEvents,
-            options?: EventOptions | undefined,
-            eventData?: PlausibleOptions | undefined,
-        ) => {
-            if (plausibleRef.current?.trackEvent) {
-                plausibleRef.current.trackEvent(eventName, options, eventData);
-            } else {
-                if (options?.callback) {
-                    options.callback();
-                }
-            }
+        (eventName: CustomEvents, options?: TrackEventOptions) => {
+            plausibleRef.current?.trackEvent(eventName, options);
         },
         [],
     );
